created abstract class
refactored classes in components
