+++ src/view/trip-controls.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import AbstractView from "./abstract.js";
 
 const createTripControlsTemplate = () => {
   return (
@@ -9,24 +9,8 @@ const createTripControlsTemplate = () => {
   );
 };
 
-export default class TripControls {
-  constructor() {
-    this._element = null;
-  }
-
+export default class TripControls extends AbstractView {
   get template() {
     return createTripControlsTemplate();
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this.template);
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
